- Moved scheduler and it's support functions to the DL/UL classes due to specialization - Added a round of SINR estimation after UE allocation in scheduler
