+++ sharc/simulation_imt_vale.py
@@ -11,7 +11,6 @@ import numpy as np
 import math
 import sys
 import matplotlib.pyplot as plt
-from scipy.interpolate import interp1d
 
 from sharc.support.enumerations import StationType
 from sharc.topology.topology_factory import TopologyFactory
@@ -188,298 +187,12 @@ class SimulationImtVale(ABC, Observable):
 
                 self.ue.spectral_mask[ue].set_mask()
 
+    @abstractmethod
     def scheduler(self):
-
-        if self.parameters.general.imt_link == "DOWNLINK":
-            # initializing the variables total_associated_ues and num_allocated_ues
-            total_associated_ues = 0
-            num_allocated_ues = 0
-
-            bs_active = np.where(self.bs.active)[0]
-
-            for bs in bs_active:
-                ue_list = self.link[bs]
-
-                # counting the total number of UEs associated to active BSs
-                total_associated_ues += len(ue_list)
-
-                # estimating the SINR for the first allocation
-                sinr_vector = self.estimate_dl_sinr(bs, ue_list)
-
-                # sort self.link[bs] in descending order of SINR
-                self.link[bs] = [x for _, x in sorted(zip(sinr_vector, self.link[bs]))][::-1]
-                self.ue.sinr[self.link[bs]] = np.sort(sinr_vector)[::-1]
-
-                # number of available RB for the current BS
-                num_available_rbs = math.ceil(self.parameters.imt.bs_rb_load_factor*self.num_rb_per_bs[bs])
-
-                allocated_ues = list()
-
-                # allocation through the ue_list
-                for ue in self.link[bs]:
-
-                    # get the throughput per RB for the current UE
-                    ue_tput = self.get_throughput_dl(self.ue.sinr[ue])*1e3
-
-                    # calculate the number of RB required for the current UE
-                    ue_num_rb = math.ceil(self.parameters.imt.min_ue_data_rate/ue_tput)
-
-                    # checking if there are still available RBs in the current BS
-                    if num_available_rbs > ue_num_rb:
-                        # allocates the UE
-                        num_available_rbs = num_available_rbs - ue_num_rb
-                        allocated_ues.append(ue)
-
-                        # calculating the UE's bandwidth
-                        self.ue.bandwidth[ue] = ue_num_rb * self.parameters.imt.rb_bandwidth
-                        # number of RB allocated for the UE
-                        self.num_rb_per_ue[ue] = ue_num_rb
-
-                    else:
-                        # storing the x and y coordinates of the UE in outage
-                        self.get_outage_positions(self.ue.x[ue], self.ue.y[ue])
-
-                # self.link only with the allocated UEs
-                self.link[bs] = allocated_ues
-
-                # counting the number of allocated UEs on the given BS
-                num_allocated_ues += len(allocated_ues)
-
-                # activating the allocated UEs
-                self.ue.active[self.link[bs]] = np.ones(len(self.link[bs]), dtype=bool)
-
-                # calculating the BS's bandwidth
-                self.bs.bandwidth[bs] = self.num_rb_per_bs[bs] * self.parameters.imt.rb_bandwidth
-
-            if total_associated_ues != 0:
-                # calculating the outage for the current drop
-                self.outage_per_drop = 1 - num_allocated_ues/total_associated_ues
-            else:
-                self.outage_per_drop = 0
-
-        else:
-            # initializing the variables total_associated_ues and num_allocated_ues
-            total_associated_ues = 0
-            num_allocated_ues = 0
-
-            bs_active = np.where(self.bs.active)[0]
-
-            for bs in bs_active:
-                ue_list = self.link[bs]
-
-                # counting the total number of UEs associated to active BSs
-                total_associated_ues += len(ue_list)
-
-                # estimating the SINR for the first allocation
-                sinr_vector = self.estimate_ul_sinr(bs, ue_list)
-
-                # sort self.link[bs] in descending order of SINR
-                self.link[bs] = [x for _, x in sorted(zip(sinr_vector, self.link[bs]))][::-1]
-                self.ue.sinr[self.link[bs]] = np.sort(sinr_vector)[::-1]
-
-                # number of available RB for the current BS
-                num_available_rbs = math.ceil(self.parameters.imt.bs_rb_load_factor * self.num_rb_per_bs[bs])
-
-                allocated_ues = list()
-                # allocation through the ue_list
-                for ue in self.link[bs]:
-
-                    # get the throughput per RB for the current UE
-                    ue_tput = self.get_throughput_ul(self.ue.sinr[ue])*1e3
-
-                    # calculate the number of RB required for the current UE
-                    ue_num_rb = math.ceil(self.parameters.imt.min_ue_data_rate / ue_tput)
-
-                    # checking if there are still available RBs in the current BS
-                    if num_available_rbs > ue_num_rb:
-                        # allocates the UE
-                        num_available_rbs = num_available_rbs - ue_num_rb
-                        allocated_ues.append(ue)
-
-                        # calculating the UE's bandwidth
-                        self.ue.bandwidth[ue] = ue_num_rb * self.parameters.imt.rb_bandwidth
-                        # number of RB allocated for the UE
-                        self.num_rb_per_ue[ue] = ue_num_rb
-                    else:
-                        # storing the x and y coordinates of the UE in outage
-                        self.get_outage_positions(self.ue.x[ue], self.ue.y[ue])
-
-                # self.link only with the allocated UEs
-                self.link[bs] = allocated_ues
-
-                # counting the number of allocated UEs on the given BS
-                num_allocated_ues += len(allocated_ues)
-
-                # activating the allocated UEs
-                self.ue.active[self.link[bs]] = np.ones(len(self.link[bs]), dtype=bool)
-
-                # calculating the BS's bandwidth
-                self.bs.bandwidth[bs] = self.num_rb_per_bs[bs] * self.parameters.imt.rb_bandwidth
-
-                if total_associated_ues != 0:
-                    # calculating the outage for the current drop
-                    self.outage_per_drop = 1 - num_allocated_ues / total_associated_ues
-                else:
-                    self.outage_per_drop = 0
-
-    def estimate_dl_sinr(self, current_bs, ue_list):
         """
-        This method estimates the DL SINR for the first allocation done by the scheduler
+        Implements the scheduler algotithm
         """
-        # initializing the interference variable
-        rx_interference = -500 * np.ones(len(ue_list))
-
-        # determining the TX power of each BS
-        total_power = self.parameters.imt.bs_conducted_power + self.bs_power_gain
-
-        bs_active = np.where(self.bs.active)[0]
-        tx_power = dict([(bs, list()) for bs in bs_active])
-        bs_tx_power = dict([(bs, list()) for bs in bs_active])
-        for bs in bs_active:
-            # TX power equally divided among the RBs
-            tx_power[bs] = total_power[bs] - 10 * np.log10(self.num_rb_per_bs[bs])
-            # TX power to each UE connected to the current BS
-            bs_tx_power[bs] = tx_power[bs] * np.ones(len(self.link[bs]))
-
-        # array with the received power for each UE connected to the current BS
-        ue_rx_power = bs_tx_power[current_bs] \
-                        - self.parameters.imt.bs_ohmic_loss \
-                        - self.parameters.imt.ue_body_loss \
-                        - self.parameters.imt.ue_ohmic_loss \
-                        - self.coupling_loss_imt[current_bs, ue_list]
-
-        # create a list with base stations that generate interference in ue_list
-        bs_interf = [b for b in bs_active if b not in [current_bs]]
-        # eliminating BSs that don't have associated UEs
-        bs_interf = [b for b in bs_interf if self.link[b]]
-
-        # calculate intra system interference
-        for bi in bs_interf:
-            interference = bs_tx_power[bi][0] - self.parameters.imt.bs_ohmic_loss \
-                            - self.coupling_loss_imt[bi, ue_list] \
-                            - self.parameters.imt.ue_body_loss - self.parameters.imt.ue_ohmic_loss
-
-            rx_interference = 10*np.log10( \
-                    np.power(10, 0.1*rx_interference)
-                    + np.power(10, 0.1*interference))
-
-        # thermal noise (per RB)
-        thermal_noise = \
-            10 * np.log10(self.parameters.imt.BOLTZMANN_CONSTANT * self.parameters.imt.noise_temperature * 1e3) + \
-            10 * np.log10(self.parameters.imt.rb_bandwidth * 1e6) + \
-            self.ue.noise_figure[ue_list]
-
-        # total interference per UE
-        total_interference = 10 * np.log10(np.power(10, 0.1*rx_interference) + np.power(10, 0.1*thermal_noise))
-
-        # calculating the SINR for each UE
-        sinr_vector = ue_rx_power - total_interference
-
-        return sinr_vector
-
-    def estimate_ul_sinr(self, current_bs, ue_list):
-        """
-        This method estimates the DL SINR for the first allocation done by the scheduler
-        """
-        # initializing the interference variable
-        rx_interference = -500 * np.ones(len(ue_list))
-
-        # determining the TX power of each UE (no power control considered)
-        ue_tx_power = self.parameters.imt.ue_p_cmax * np.ones(len(ue_list))
-
-        # array with the received power of the BS from each UE
-        bs_rx_power = ue_tx_power  \
-                     - self.parameters.imt.ue_ohmic_loss \
-                     - self.parameters.imt.ue_body_loss \
-                     - self.coupling_loss_imt[current_bs, ue_list] \
-                     - self.parameters.imt.bs_ohmic_loss
-
-        # create a list with base stations that generate interference in the current BS
-        bs_active = np.where(self.bs.active)[0]
-        bs_interf = [b for b in bs_active if b not in [current_bs]]
-        # eliminating BSs that don't have associated UEs
-        bs_interf = [b for b in bs_interf if self.link[b]]
-
-        # calculate intra system interference (consider only the UE closest to the current BS)
-        for bi in bs_interf:
-            ui = self.link[bi]
-
-            # coupling loss from the interfering UEs to the current BS
-            ui_losses = self.coupling_loss_imt[current_bs, ui]
-
-            # minimum coupling loss between the current BS and an adjacent UE
-            min_ui_loss = min(ui_losses)
-
-            interference = ue_tx_power - self.parameters.imt.ue_ohmic_loss \
-                           - self.parameters.imt.ue_body_loss \
-                           - min_ui_loss - self.parameters.imt.bs_ohmic_loss
-
-            rx_interference = 10*np.log10( \
-                    np.power(10, 0.1*rx_interference)
-                    + np.power(10, 0.1*interference))
-
-        # calculate N
-        thermal_noise = \
-            10 * np.log10(self.parameters.imt.BOLTZMANN_CONSTANT * self.parameters.imt.noise_temperature * 1e3) + \
-            10 * np.log10(self.parameters.imt.rb_bandwidth * self.num_rb_per_bs[current_bs] * 1e6) + \
-            self.bs.noise_figure[current_bs]
-
-        # total interference per UE
-        total_interference = 10 * np.log10(np.power(10, 0.1 * rx_interference) + np.power(10, 0.1 * thermal_noise))
-
-        # calculating the SINR for each UE
-        sinr_vector = bs_rx_power - total_interference
-
-        return sinr_vector
-
-    def get_throughput_dl(self, ue_sinr):
-        """
-        This method returns the throughput per RB in kbps for a given SINR in the DL
-        """
-        sinr_vals = [-6, -5, -4, -3, -2, -1, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20,
-                     21, 22, 23, 24, 25]
-        tput_vals = [39.7154, 48.2521, 58.2422, 69.7951, 82.9873, 97.8559, 114.3942, 132.5529, 152.2445, 173.3518,
-                     195.7379, 219.2562, 243.7591, 269.1051, 295.1634, 321.8168, 348.9628, 376.5132, 404.5830, 442.8072,
-                     481.8425, 521.5469, 561.7995, 602.4986, 643.5594, 684.9121, 726.4998, 768.2760, 810.2032, 852.2511,
-                     894.3953, 936.6164]
-
-        if ue_sinr > 25:
-            ue_tput = 936.6164
-        elif ue_sinr < - 6:
-            # negligible value to represent a throughput equal to zero. Cannot set to zero due to the division done in
-            # the scheduler method
-            ue_tput = np.power(0.1, 50)
-        else:
-            # interpolate and generate MCS curve
-            mcs_curve = interp1d(sinr_vals, tput_vals)
-
-            # throughput per RB for the given MCS
-            ue_tput = mcs_curve(ue_sinr)
-
-        return ue_tput
-
-    def get_throughput_ul(self, ue_sinr):
-        """
-        This method returns the throughput per RB in kbps for a given SINR in the UL
-        """
-        sinr_vals = [-11.5, -10, -6, -2, 2, 6, 10, 14, 18, 22, 24.1, 30]
-        tput_vals = [4.492, 8.532, 29.956, 71.008, 126.814, 211.917, 322.393, 443.439, 575.342, 692.987, 720.702,
-                     720.702]
-
-        if ue_sinr > 30:
-            ue_tput = 720.702
-        elif ue_sinr < - 11.5:
-            # negligible value to represent a throughput equal to zero. Cannot set to zero due to the division done in
-            # the scheduler method
-            ue_tput = np.power(0.1, 50)
-        else:
-            # interpolate and generate MCS curve
-            mcs_curve = interp1d(sinr_vals, tput_vals)
-
-            # throughput per RB for the given MCS
-            ue_tput = mcs_curve(ue_sinr)
-
-        return ue_tput
+        pass
 
     def get_outage_positions(self, ue_x, ue_y):
         """
@@ -500,7 +213,6 @@ class SimulationImtVale(ABC, Observable):
             index = self.ues_in_outage_coordinates.index(outage_position)
             # counts one more occurrence
             self.ues_in_outage_counter[index] += 1
-
 
     def calculate_gains(self, station_1: StationManager, station_2: StationManager) -> np.array:
         """
